feat: Refactor websocket discussion create listener
