+++ src/services/api/websocket/discussion/index.js
@@ -1,0 +1,5 @@
+import create from './create';
+
+export default {
+  create,
+};
