+++ src/services/api/websocket/index.js
@@ -1,7 +1,9 @@
 import room from './room';
+import discussion from './discussion';
 import soundNotification from './soundNotification';
 
 export default {
   room,
+  discussion,
   soundNotification,
 };
